requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
numpy==1.24.0
pydantic==2.4.2
ffmpeg-python==0.2.0
psutil==5.9.5
python-json-logger==2.0.7
websockets==12.0
requests==2.31.0  # For testing
pytest==7.4.3  # For testing
pytest-asyncio==0.21.1  # For testing
